Recursive Quicksort
Quicksort with recursive structure. It would exceed the maximum recursion depth if the size of the input is more than n=600. It's suitable for small size lists. The expected (average) time complexity of this algorithm for a list of size "n" is "c n log(n)" where "c" is a constant.